Split structure specs and geometry helpers
